# inference: remove `throw` block deoptimization completely