@nanosoldier 2023-04-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49260#issuecomment-1497455006) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/9b699c1_vs_bd1a664/report.md).