@oscardssmith 2023-04-20
I think we should merge this. Now that tab complete relies on proving termination and effectfree-ness, this PR also makes tab complete better.