@nanosoldier 2023-09-20
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49260#issuecomment-1727446334) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/2105313_vs_4923e95/report.md).