@aviatesk 2024-01-18
@nanosoldier `runbenchmarks("inference", vs=":master")`