@nanosoldier 2024-01-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49260#issuecomment-1897950636) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/220b6f2_vs_8f2f178/report.md).