@aviatesk 2024-04-11
Ok. While there are a few clear regressions, I believe the downsides in the other compiler applications outweigh them. I've been wanting to get rid of the special cases for this functionality in JET too.