@aviatesk 2024-04-11
@nanosoldier `runbenchmarks("inference", vs=":master")`