@nanosoldier 2024-04-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49260#issuecomment-2049759725) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/6e1a1ac_vs_a7fd6a7/report.md).