@aviatesk 2024-04-12
It looks like the regressions in cases such as `["inference", "allinference", "rand(Float64)"]` are indeed real. Should we go ahead regardless?