@gbaraldi 2024-04-12
I think we can go ahead!