@aviatesk 2024-04-12
I've done a deeper dive, and it looks like the regression is specifically caused by this code: https://github.com/JuliaLang/julia/blob/be3bc9af09deb28e51c06854278ce4de099f500d/base/error.jl#L95
Here’s what I found:
```julia
julia> @newinterp LocalCacheInterp

julia> l95(b::Bool, tag::UInt, i::Int) = Base.inferencebarrier(b) ? throw(ArgumentError("Unexpected extended backtrace entry tag $tag at bt[$i]")) : nothing;

julia> interp = LocalCacheInterp();

julia> Base.infer_return_type(l95, (Bool,UInt,Int); interp)
Nothing

julia> length(interp.code_cache.dict)
2 # on master
593 # this PR
```
(For context, running `Base.infer_return_type(rand, (Type{Float64},); interp)` results in about 700 entries in `interp.code_cache.dict`, pointing out that string interpolation is a major part of the regression).

Considering this, it might be sensible to add a special handling for the pattern of `throw(... something with string interpolation...)`.