@vtjnash 2024-04-12
Could we make that a LazyString inside the Exception object and teach Test to treat those as equivalent so that not much breaks?