@aviatesk 2024-04-15
Are you proposing an optimization where the optimizer replaces `string(...)::String` with `lazystring(...)::LazyString` when the return value is passed into an `<:Exception` object? What would be the approach for scenarios like this?
```julia
struct MyException <: Exception
    msg::String
end

let x = ...
    throw(MyException("... $x ..."))
end
```