@aviatesk 2024-04-15
I thought we could just restrict the scope of `unoptimize_throw_blocks` to the `string(...)` calls that are known to be `throw`n.