@gbaraldi 2024-04-24
So talking with @topolarity I think that benchmark result is a bit misleading because sure, it's a lot of methods that get inferred, but they would only be new in an empty inference cache. From a top level glance they are mostly normal integer operations and any of amount of string printing (which is very likely to already be inferred) so maybe the cost isn't too large on a "real" world situation.