@topolarity 2024-04-24
As an example:
```julia
julia> @newinterp LocalCacheInterp true

julia> l95(b::Bool, tag::UInt, i::Int) = Base.inferencebarrier(b) ? throw(ArgumentError("Unexpected extended backtrace entry tag $tag at bt[$i]")) : nothing;

julia> make_err(b::Bool, name::String, x::Float64) = ArgumentError("Found interesting $name with value $x? $b")

julia> interp = LocalCacheInterp();

julia> Base.infer_return_type(make_err, (Bool, String, Float64); interp); length(interp.code_cache.dict)
850

julia> Base.infer_return_type(l95, (Bool,UInt,Int); interp); length(interp.code_cache.dict)
921
```

The inference of `l95(...)` contributed ~71 new code instances, but that's not terribly surprising given that it is interpolating a different set of types into a string.

Maybe it'd be more profitable to make our string interpolation code less eager to generate so many specializations?