@aviatesk 2024-08-07
@nanosoldier `runbenchmarks("inference", vs=":master")`