@nanosoldier 2024-08-07
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49260#issuecomment-2274119325) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/046452c_vs_bd582f7/report.md).