@nsajko 2024-08-25
This causes a weird inference regression, see #55583.